refactor: 도메인 이벤트 클래스를 common 패키지로 이동하고 각 모듈 ModuleInfo에서 불필요한 common 명시를 제거
