+++ tamago-core/src/main/kotlin/tamago/server/core/letter/ModuleInfo.kt
@@ -3,6 +3,6 @@ package tamago.server.core.letter
 import org.springframework.modulith.ApplicationModule
 import org.springframework.modulith.PackageInfo
 
-@ApplicationModule(allowedDependencies = ["common", "running", "notification"])
+@ApplicationModule(allowedDependencies = ["running", "notification"])
 @PackageInfo
 class LetterModuleInfo
