+++ tamago-core/src/main/kotlin/tamago/server/core/notification/ModuleInfo.kt
@@ -3,6 +3,6 @@ package tamago.server.core.notification
 import org.springframework.modulith.ApplicationModule
 import org.springframework.modulith.PackageInfo
 
-@ApplicationModule(allowedDependencies = ["common"])
+@ApplicationModule
 @PackageInfo
 class NotificationModuleInfo
