+++ tamago-core/src/main/kotlin/tamago/server/core/refreshtoken/ModuleInfo.kt
@@ -3,6 +3,6 @@ package tamago.server.core.refreshtoken
 import org.springframework.modulith.ApplicationModule
 import org.springframework.modulith.PackageInfo
 
-@ApplicationModule(allowedDependencies = ["common"])
+@ApplicationModule
 @PackageInfo
 class RefreshTokenModuleInfo
