+++ tamago-core/src/main/kotlin/tamago/server/core/running/ModuleInfo.kt
@@ -3,6 +3,6 @@ package tamago.server.core.running
 import org.springframework.modulith.ApplicationModule
 import org.springframework.modulith.PackageInfo
 
-@ApplicationModule(allowedDependencies = ["common", "monster"])
+@ApplicationModule(allowedDependencies = ["monster"])
 @PackageInfo
 class RunningModuleInfo
