+++ tamago-core/src/main/kotlin/tamago/server/core/user/ModuleInfo.kt
@@ -3,6 +3,6 @@ package tamago.server.core.user
 import org.springframework.modulith.ApplicationModule
 import org.springframework.modulith.PackageInfo
 
-@ApplicationModule(allowedDependencies = ["common", "refreshtoken"])
+@ApplicationModule(allowedDependencies = ["refreshtoken"])
 @PackageInfo
 class UserModuleInfo
