+++ tamago-core/src/main/kotlin/tamago/server/core/user/application/service/UserCommandService.kt
@@ -3,6 +3,7 @@ package tamago.server.core.user.application.service
 import org.springframework.context.ApplicationEventPublisher
 import org.springframework.security.crypto.password.PasswordEncoder
 import org.springframework.stereotype.Service
+import tamago.server.core.common.event.UserSignedUpEvent
 import tamago.server.core.common.jwt.JwtTokenProvider
 import tamago.server.core.common.vo.UserId
 import tamago.server.core.refreshtoken.RefreshTokenCommandUseCase
@@ -10,7 +11,6 @@ import tamago.server.core.user.UserCommandUseCase
 import tamago.server.core.user.application.exception.UserSaveErrorException
 import tamago.server.core.user.domain.aggregate.User
 import tamago.server.core.user.domain.enum.AuthProvider
-import tamago.server.core.user.domain.event.UserSignedUpEvent
 import tamago.server.core.user.domain.port.inbound.command.LoginCommandDto
 import tamago.server.core.user.domain.port.inbound.command.SignUpCommandDto
 import tamago.server.core.user.domain.port.inbound.query.TokenQueryDto
